Move MeasurementValue class
